Add .prettierrc for vscode built-in format on save
Astro prettier plugin integration wants to use a .prettierrc.mjs, but the
.mjs file does not play well with vscode's built-in format on save
feature, which seems to prefer having a .prettierrc. So add a .prettierrc
and have duplicate rules with those in the .prettierrc.mjs. The former
is for vscode to format on save and the latter is for running
`npx prettier src -w` via the CLI.